refactor: date picker dialog
